add more assets and create boot scene
